Highscores upadte

+++ src/main/java/com/Hopur26/D/shooter/Services/HighScoreService.java
@@ -6,5 +6,5 @@ import com.Hopur26.D.shooter.Persistence.Entities.News;
 public interface HighScoreService {
     HighScore save(HighScore highScore);
     HighScore create(Integer gameID, String gamename);
-    HighScore findByID(long ID);
+    HighScore findByID(Integer ID);
 }
