+++ src/main/java/com/Hopur26/D/shooter/Services/Implementation/HighScoreServiceImplementation.java
@@ -34,7 +34,7 @@ public class HighScoreServiceImplementation implements HighScoreService {
 
 
     @Override
-    public HighScore findByID(long ID) {
+    public HighScore findByID(Integer ID) {
         return highScoreRepository.findByID(ID);
     }
